update
updated db
updated DataClass(updated ParseRegNum, added GetInsurerINN, GetInsurerShortName)
added SZVMPage
removed middle colon in MainPage
removed some old code and reference, and optimize some moments

+++ source/Model/Insurer.cs
@@ -21,6 +21,6 @@ namespace Source.Model
         public string FirstName { get; set; }
         public string MiddleName { get; set; }
         public string InsuranceNumber { get; set; }
-        public Nullable<int> INN { get; set; }
+        public Nullable<long> INN { get; set; }
     }
 }
